+++ src/aks-preview/azext_aks_preview/_params.py
@@ -14,6 +14,20 @@ from azext_aks_preview._completers import (
     get_k8s_versions_completion_list,
     get_vm_size_completion_list,
 )
+from azure.cli.command_modules.acs._consts import (
+    CONST_OUTBOUND_TYPE_LOAD_BALANCER,
+    CONST_OUTBOUND_TYPE_MANAGED_NAT_GATEWAY,
+    CONST_OUTBOUND_TYPE_USER_ASSIGNED_NAT_GATEWAY,
+    CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING,
+)
+from azure.cli.command_modules.acs._validators import (
+    validate_load_balancer_idle_timeout,
+    validate_load_balancer_outbound_ip_prefixes,
+    validate_load_balancer_outbound_ips,
+    validate_load_balancer_outbound_ports,
+    validate_nat_gateway_idle_timeout,
+    validate_nat_gateway_managed_outbound_ip_count,
+)
 from azext_aks_preview._consts import (
     CONST_ABSOLUTEMONTHLY_MAINTENANCE_SCHEDULE,
     CONST_AZURE_KEYVAULT_NETWORK_ACCESS_PRIVATE,
@@ -58,10 +72,6 @@ from azext_aks_preview._consts import (
     CONST_OS_SKU_UBUNTU,
     CONST_OS_SKU_WINDOWS2019,
     CONST_OS_SKU_WINDOWS2022,
-    CONST_OUTBOUND_TYPE_LOAD_BALANCER,
-    CONST_OUTBOUND_TYPE_MANAGED_NAT_GATEWAY,
-    CONST_OUTBOUND_TYPE_USER_ASSIGNED_NAT_GATEWAY,
-    CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING,
     CONST_PATCH_UPGRADE_CHANNEL,
     CONST_RAPID_UPGRADE_CHANNEL,
     CONST_RELATIVEMONTHLY_MAINTENANCE_SCHEDULE,
@@ -117,15 +127,9 @@ from azext_aks_preview._validators import (
     validate_k8s_version,
     validate_linux_host_name,
     validate_load_balancer_backend_pool_type,
-    validate_load_balancer_idle_timeout,
-    validate_load_balancer_outbound_ip_prefixes,
-    validate_load_balancer_outbound_ips,
-    validate_load_balancer_outbound_ports,
     validate_load_balancer_sku,
     validate_max_surge,
     validate_message_of_the_day,
-    validate_nat_gateway_idle_timeout,
-    validate_nat_gateway_managed_outbound_ip_count,
     validate_node_public_ip_tags,
     validate_nodepool_id,
     validate_nodepool_labels,