+++ src/aks-preview/HISTORY.rst
@@ -9,8 +9,10 @@ If there is no rush to release a new version, please just add a description of t
 
 To release a new version, please select a new version number (usually plus 1 to last patch version, X.Y.Z -> Major.Minor.Patch, more details in `\doc <https://semver.org/>`_), and then add a new section named as the new version number in this file, the content should include the new modifications and everything from the *Pending* section. Finally, update the `VERSION` variable in `setup.py` with this new version number.
 
-Pending
+0.5.152
 ++++++
+* move loadbalancer/natgateway util functions to azure-cli and update reference in aks-preview project. 
+* bump azure-cli to 2.49
 
 0.5.151
 +++++++